start laravel9
Hi this is my first project which i admit it is very challenging for me as i never write framework based code in laravel 9. Personally I find it difficult to install different libraries when installing it. It looks very easy to look very fast in writing this project, but I, when I try to write it, it is very difficult to follow the file that I didn't create. Finally, it took a little time to study. and create new works based on the original Thank you for coming to read this.